+++ src/app/beer_garden/router.py
@@ -307,7 +307,7 @@ def update_api_heartbeat(operation: Operation):
                 local_garden = get_garden(config.get("garden.name"))
 
                 # Will only support mapping 1 hop away legacy Garden Syncs
-                child_garden = False
+                multi_hop_garden = True
                 for child in local_garden.children:
                     if child.name == operation.model.payload.name:
                         logger.warning(
@@ -317,9 +317,9 @@ def update_api_heartbeat(operation: Operation):
                             )
                         )
                         operation.source_garden_name = operation.model.payload.name
-                        child_garden = True
+                        multi_hop_garden = False
                         break
-                if child_garden:
+                if multi_hop_garden:
                     return
             else:
                 return